requirements.txt
# Core Requirements
torch==2.0.0
torch-geometric==2.3.0
numpy==1.23.5
scikit-learn==1.2.0
scipy==1.10.0
pandas==1.5.3

# Text Processing
chardet==5.1.0
tqdm==4.65.0
sentencepiece==0.1.99
nltk==3.8.1

# Graph ML Dependencies
torch-scatter==2.1.0
torch-sparse==0.6.16
torch-cluster==1.6.0
torch-spline-conv==1.2.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Data Serialization
pickle-mixin==1.0.2
jsonpickle==3.0.2
pyyaml==6.0.1

# Jupyter Support (for notebook)
jupyter==1.0.0
ipython==8.12.0

# Additional Utilities
protobuf==3.20.3
loguru==0.7.0
wandb==0.15.0  # Optional for experiment tracking
python-dotenv==1.0.0  # For environment management

# Special Cases (explicit versions for compatibility)
networkx==3.1  # Required by torch-geometric
pynvml==11.5.0  # For GPU monitoring